+++ src/auth.module.ts
@@ -1,12 +1,14 @@
 import { NgModule } from '@angular/core';
 import { HTTP_INTERCEPTORS } from '@angular/common/http';
 
-import { PublicGuard } from './public.guard';
-import { ProtectedGuard } from './protected.guard';
+import { AuthGuard } from './guards/auth.guard';
+import { PublicGuard } from './guards/public.guard';
+import { ProtectedGuard } from './guards/protected.guard';
 import { AuthInterceptor } from './auth.interceptor';
 
 @NgModule({
   providers: [
+    AuthGuard,
     PublicGuard,
     ProtectedGuard,
     AuthInterceptor,
