+++ src/guards/auth.guard.ts
@@ -1,29 +1,29 @@
 import { Injectable, Inject } from '@angular/core';
-import { CanActivate, CanActivateChild, Router, ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
+import { Router, CanActivate, CanActivateChild, ActivatedRouteSnapshot } from '@angular/router';
 import { Observable } from 'rxjs/Observable';
 
-import { AuthService } from './auth.service';
-import { AUTH_SERVICE, PROTECTED_FALLBACK_PAGE_URI } from './tokens';
+import { AuthService } from '../auth.service';
+import { AUTH_SERVICE, LOGIN_PAGE_URI } from '../tokens';
 
 /**
  * Guard, checks access token availability and allows or disallows access to page,
  * and redirects out
  *
- * usage: { path: 'test', component: TestComponent, canActivate: [ PublicGuard ] }
+ * usage: { path: 'test', component: TestComponent, canActivate: [ AuthGuard ] }
  *
  * @export
  *
- * @class PublicGuard
+ * @class AuthGuard
  *
  * @implements {CanActivate}
  * @implements {CanActivateChild}
  */
 @Injectable()
-export class PublicGuard implements CanActivate, CanActivateChild {
+export class AuthGuard implements CanActivate, CanActivateChild {
 
   constructor(
     @Inject(AUTH_SERVICE) private authService: AuthService,
-    @Inject(PROTECTED_FALLBACK_PAGE_URI) private protectedFallbackPageUri: string,
+    @Inject(LOGIN_PAGE_URI) private loginPageUri: string,
     private router: Router
   ) { }
 
@@ -35,11 +35,11 @@ export class PublicGuard implements CanActivate, CanActivateChild {
    *
    * @returns {Observable<boolean>}
    */
-  public canActivate(_route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<boolean> {
+  public canActivate(_route: ActivatedRouteSnapshot): Observable<boolean> {
     return this.authService.isAuthorized()
       .map(isAuthorized => {
-        if (isAuthorized && !this.isProtectedPage(state)) {
-          this.navigate(this.protectedFallbackPageUri);
+        if (!isAuthorized) {
+          this.navigate(this.loginPageUri);
           return false;
         }
         return true;
@@ -54,21 +54,8 @@ export class PublicGuard implements CanActivate, CanActivateChild {
    *
    * @returns {Observable<boolean>}
    */
-  public canActivateChild(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<boolean> {
-    return this.canActivate(route, state);
-  }
-
-  /**
-   * Check, if current page is protected fallback page
-   *
-   * @private
-   *
-   * @param {RouterStateSnapshot} state
-   *
-   * @returns {boolean}
-   */
-  private isProtectedPage(state: RouterStateSnapshot): boolean {
-    return state.url === this.protectedFallbackPageUri;
+  public canActivateChild(route: ActivatedRouteSnapshot): Observable<boolean> {
+    return this.canActivate(route);
   }
 
   /**
