+++ src/guards/protected.guard.ts
@@ -2,8 +2,8 @@ import { Injectable, Inject } from '@angular/core';
 import { Router, CanActivate, CanActivateChild, ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
 import { Observable } from 'rxjs/Observable';
 
-import { AuthService } from './auth.service';
-import { AUTH_SERVICE, PUBLIC_FALLBACK_PAGE_URI } from './tokens';
+import { AuthService } from '../auth.service';
+import { AUTH_SERVICE, PUBLIC_FALLBACK_PAGE_URI } from '../tokens';
 
 /**
  * Guard, checks access token availability and allows or disallows access to page,
