+++ src/index.ts
@@ -1,5 +1,6 @@
 export { AuthService } from './auth.service';
-export { PublicGuard } from './public.guard';
-export { ProtectedGuard } from './protected.guard';
-export { AUTH_SERVICE, PUBLIC_FALLBACK_PAGE_URI, PROTECTED_FALLBACK_PAGE_URI } from './tokens';
+export { AuthGuard } from './guards/auth.guard';
+export { PublicGuard } from './guards/public.guard';
+export { ProtectedGuard } from './guards/protected.guard';
+export { AUTH_SERVICE, LOGIN_PAGE_URI, PUBLIC_FALLBACK_PAGE_URI, PROTECTED_FALLBACK_PAGE_URI } from './tokens';
 export { AuthModule } from './auth.module';
